feat: update mailer to use Symfony Mailer, add MAIL_DSN configuration, and clean up code

+++ src/AbraFlexi/Mailer/DocumentMailer.php
@@ -319,7 +319,9 @@ class DocumentMailer extends HtmlMailer
         $result = parent::send();
 
         foreach ($this->cleanup as $tmp) {
-            unlink($tmp);
+            if ($tmp) {
+                unlink($tmp);
+            }
         }
 
         return $result;
